Extract budget functions from gateway_api.py into gateway_budget.py
Move all budget-related functions into a dedicated module:
- backfill_budget_spend, maybe_refresh_budget_policies,
  maybe_record_budget_cost, fire_budget_exceeded_webhooks,
  get_model_info, make_cost_recording_reducer

Also fix incorrect type/action names:
- BudgetPolicyCrossedPayload → BudgetPolicyExceededPayload
- WebhookAction.CROSSED → WebhookAction.EXCEEDED
