+++ mlflow/server/gateway_api.py
@@ -15,11 +15,8 @@ from typing import Any
 from fastapi import APIRouter, HTTPException, Request
 from fastapi.responses import StreamingResponse
 
-from mlflow.entities.gateway_budget_policy import BudgetTargetScope
 from mlflow.entities.gateway_endpoint import GatewayModelLinkageType
-from mlflow.entities.webhook import WebhookAction, WebhookEntity, WebhookEvent
 from mlflow.exceptions import MlflowException
-from mlflow.gateway.budget_tracker import BudgetWindow, get_budget_tracker
 from mlflow.gateway.config import (
     AnthropicConfig,
     EndpointConfig,
@@ -42,10 +39,14 @@ from mlflow.gateway.providers.base import (
     TrafficRouteProvider,
 )
 from mlflow.gateway.schemas import chat, embeddings
-from mlflow.gateway.tracing_utils import aggregate_chat_stream_chunks, maybe_traced_gateway_call
+from mlflow.gateway.tracing_utils import maybe_traced_gateway_call
 from mlflow.gateway.utils import safe_stream, to_sse_chunk, translate_http_exception
 from mlflow.protos.databricks_pb2 import RESOURCE_DOES_NOT_EXIST
-from mlflow.server.handlers import _get_model_registry_store
+from mlflow.server.gateway_budget import (
+    get_model_info,
+    make_cost_recording_reducer,
+    maybe_record_budget_cost,
+)
 from mlflow.store.tracking.abstract_store import AbstractStore
 from mlflow.store.tracking.gateway.config_resolver import get_endpoint_config
 from mlflow.store.tracking.gateway.entities import (
@@ -56,12 +57,9 @@ from mlflow.store.tracking.gateway.entities import (
 from mlflow.store.tracking.sqlalchemy_store import SqlAlchemyStore
 from mlflow.telemetry.events import GatewayInvocationEvent, GatewayInvocationType
 from mlflow.telemetry.track import _record_event
-from mlflow.tracing.constant import CostKey, TokenUsageKey, TraceMetadataKey
-from mlflow.tracing.utils import calculate_cost_by_model_and_token_usage
+from mlflow.tracing.constant import TraceMetadataKey
 from mlflow.tracking._tracking_service.utils import _get_store
 from mlflow.utils.workspace_context import get_request_workspace
-from mlflow.webhooks.delivery import deliver_webhook
-from mlflow.webhooks.types import BudgetPolicyCrossedPayload
 
 _logger = logging.getLogger(__name__)
 
@@ -395,176 +393,6 @@ def _validate_store(store: AbstractStore) -> None:
         )
 
 
-def _backfill_budget_spend(store: SqlAlchemyStore, tracker, new_windows) -> None:
-    """Backfill cumulative spend for newly created budget windows from trace history.
-
-    When a new BudgetWindow is created (server restart or new policy), its
-    cumulative_spend starts at 0. This queries historical trace cost data to
-    seed the spend so budgets aren't effectively reset on restart.
-    """
-    if not new_windows:
-        return
-
-    for window in new_windows:
-        try:
-            start_ms = int(window.window_start.timestamp() * 1000)
-            end_ms = int(window.window_end.timestamp() * 1000)
-            workspace = (
-                window.policy.workspace
-                if window.policy.target_scope == BudgetTargetScope.WORKSPACE
-                else None
-            )
-            spend = store.sum_gateway_trace_cost(
-                start_time_ms=start_ms,
-                end_time_ms=end_ms,
-                workspace=workspace,
-            )
-            if spend > 0:
-                tracker.backfill_spend(window.policy.budget_policy_id, spend)
-        except Exception:
-            _logger.debug(
-                "Failed to backfill budget spend for policy %s",
-                window.policy.budget_policy_id,
-                exc_info=True,
-            )
-
-
-def _maybe_refresh_budget_policies(store: SqlAlchemyStore) -> None:
-    """Refresh budget policies from the database if stale."""
-    tracker = get_budget_tracker()
-    if tracker.needs_refresh():
-        try:
-            policies = store.list_budget_policies()
-            new_windows = tracker.load_policies(policies)
-            _backfill_budget_spend(store, tracker, new_windows)
-        except Exception:
-            _logger.debug("Failed to refresh budget policies", exc_info=True)
-
-
-def _maybe_record_budget_cost(
-    store: SqlAlchemyStore,
-    response: Any,
-    model_name: str | None = None,
-    model_provider: str | None = None,
-    workspace: str | None = None,
-) -> None:
-    """Record cost from a gateway response against budget policies.
-
-    Extracts token usage from the response, calculates cost via LiteLLM,
-    records it in the budget tracker, and fires webhooks for newly-crossed
-    budget windows.
-
-    This is a best-effort operation; errors are logged but not raised.
-    """
-    try:
-        # Extract token usage from response
-        usage_dict = None
-        if hasattr(response, "usage") and response.usage is not None:
-            usage = response.usage
-            usage_dict = {}
-            if hasattr(usage, "prompt_tokens") and usage.prompt_tokens is not None:
-                usage_dict[TokenUsageKey.INPUT_TOKENS] = usage.prompt_tokens
-            if hasattr(usage, "completion_tokens") and usage.completion_tokens is not None:
-                usage_dict[TokenUsageKey.OUTPUT_TOKENS] = usage.completion_tokens
-            if hasattr(usage, "total_tokens") and usage.total_tokens is not None:
-                usage_dict[TokenUsageKey.TOTAL_TOKENS] = usage.total_tokens
-        elif isinstance(response, dict):
-            if raw_usage := response.get("usage"):
-                usage_dict = {
-                    TokenUsageKey.INPUT_TOKENS: raw_usage.get(
-                        "prompt_tokens", raw_usage.get("input_tokens", 0)
-                    ),
-                    TokenUsageKey.OUTPUT_TOKENS: raw_usage.get(
-                        "completion_tokens", raw_usage.get("output_tokens", 0)
-                    ),
-                }
-
-        if not usage_dict:
-            return
-
-        cost = calculate_cost_by_model_and_token_usage(
-            model_name=model_name,
-            usage=usage_dict,
-            model_provider=model_provider,
-        )
-        if not cost:
-            return
-
-        total_cost = cost.get(CostKey.TOTAL_COST, 0.0)
-        if total_cost <= 0:
-            return
-
-        _maybe_refresh_budget_policies(store)
-        tracker = get_budget_tracker()
-        # Fire webhooks for newly-crossed budget windows
-        if newly_crossed := tracker.record_cost(total_cost, workspace=workspace):
-            _fire_budget_crossed_webhooks(newly_crossed, workspace)
-    except Exception:
-        _logger.debug("Failed to record budget cost", exc_info=True)
-
-
-def _fire_budget_crossed_webhooks(newly_crossed: list[BudgetWindow], workspace: str | None) -> None:
-    """Fire budget_crossed webhooks for newly-crossed budget windows."""
-    try:
-        registry_store = _get_model_registry_store()
-    except Exception:
-        _logger.debug("Failed to get model registry store for webhook delivery", exc_info=True)
-        return
-
-    event = WebhookEvent(WebhookEntity.BUDGET_POLICY, WebhookAction.CROSSED)
-
-    for window in newly_crossed:
-        policy = window.policy
-        if policy.budget_action.value != "ALERT":
-            continue
-
-        payload = BudgetPolicyCrossedPayload(
-            budget_policy_id=policy.budget_policy_id,
-            budget_unit=policy.budget_unit.value,
-            budget_amount=policy.budget_amount,
-            current_spend=window.cumulative_spend,
-            duration_unit=policy.duration_unit.value,
-            duration_value=policy.duration_value,
-            target_scope=policy.target_scope.value,
-            workspace=workspace or (policy.workspace or "default"),
-            window_start=int(window.window_start.timestamp() * 1000),
-        )
-        deliver_webhook(event=event, payload=payload, store=registry_store)
-
-
-def _get_model_info(
-    endpoint_config: GatewayEndpointConfig,
-) -> tuple[str | None, str | None]:
-    """Extract model_name and provider from endpoint config."""
-    if endpoint_config.models:
-        m = endpoint_config.models[0]
-        return m.model_name, m.provider
-    return None, None
-
-
-def _make_cost_recording_reducer(
-    store: SqlAlchemyStore,
-    endpoint_config: GatewayEndpointConfig,
-    workspace: str | None,
-):
-    """Create an output_reducer that aggregates stream chunks and records budget cost."""
-
-    def reducer(chunks):
-        result = aggregate_chat_stream_chunks(chunks)
-        if result:
-            model_name, model_provider = _get_model_info(endpoint_config)
-            _maybe_record_budget_cost(
-                store,
-                result,
-                model_name=model_name,
-                model_provider=model_provider,
-                workspace=workspace,
-            )
-        return result
-
-    return reducer
-
-
 def _extract_endpoint_name_from_model(body: dict[str, Any]) -> str:
     """
     Extract and validate the endpoint name from the 'model' parameter in the request body.
@@ -625,7 +453,7 @@ async def invocations(endpoint_name: str, request: Request):
                 provider.chat_stream,
                 endpoint_config,
                 user_metadata,
-                output_reducer=_make_cost_recording_reducer(store, endpoint_config, workspace),
+                output_reducer=make_cost_recording_reducer(store, endpoint_config, workspace),
                 request_headers=headers,
                 request_type=GatewayRequestType.UNIFIED_CHAT,
             )(payload)
@@ -641,8 +469,8 @@ async def invocations(endpoint_name: str, request: Request):
                 request_headers=headers,
                 request_type=GatewayRequestType.UNIFIED_CHAT,
             )(payload)
-            model_name, model_provider = _get_model_info(endpoint_config)
-            _maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
+            model_name, model_provider = get_model_info(endpoint_config)
+            maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
             return response
 
     elif "input" in body:
@@ -664,8 +492,8 @@ async def invocations(endpoint_name: str, request: Request):
             request_headers=headers,
             request_type=GatewayRequestType.UNIFIED_EMBEDDINGS,
         )(payload)
-        model_name, model_provider = _get_model_info(endpoint_config)
-        _maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
+        model_name, model_provider = get_model_info(endpoint_config)
+        maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
         return response
 
     else:
@@ -720,7 +548,7 @@ async def chat_completions(request: Request):
             provider.chat_stream,
             endpoint_config,
             user_metadata,
-            output_reducer=_make_cost_recording_reducer(store, endpoint_config, workspace),
+            output_reducer=make_cost_recording_reducer(store, endpoint_config, workspace),
             request_headers=headers,
             request_type=GatewayRequestType.UNIFIED_CHAT,
         )(payload)
@@ -736,8 +564,8 @@ async def chat_completions(request: Request):
             request_headers=headers,
             request_type=GatewayRequestType.UNIFIED_CHAT,
         )(payload)
-        model_name, model_provider = _get_model_info(endpoint_config)
-        _maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
+        model_name, model_provider = get_model_info(endpoint_config)
+        maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
         return response
 
 
@@ -808,8 +636,8 @@ async def openai_passthrough_chat(request: Request):
     response = await traced_passthrough(
         action=PassthroughAction.OPENAI_CHAT, payload=body, headers=headers
     )
-    model_name, model_provider = _get_model_info(endpoint_config)
-    _maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
+    model_name, model_provider = get_model_info(endpoint_config)
+    maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
     return response
 
 
@@ -855,8 +683,8 @@ async def openai_passthrough_embeddings(request: Request):
     response = await traced_passthrough(
         action=PassthroughAction.OPENAI_EMBEDDINGS, payload=body, headers=headers
     )
-    model_name, model_provider = _get_model_info(endpoint_config)
-    _maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
+    model_name, model_provider = get_model_info(endpoint_config)
+    maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
     return response
 
 
@@ -927,8 +755,8 @@ async def openai_passthrough_responses(request: Request):
     response = await traced_passthrough(
         action=PassthroughAction.OPENAI_RESPONSES, payload=body, headers=headers
     )
-    model_name, model_provider = _get_model_info(endpoint_config)
-    _maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
+    model_name, model_provider = get_model_info(endpoint_config)
+    maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
     return response
 
 
@@ -999,8 +827,8 @@ async def anthropic_passthrough_messages(request: Request):
     response = await traced_passthrough(
         action=PassthroughAction.ANTHROPIC_MESSAGES, payload=body, headers=headers
     )
-    model_name, model_provider = _get_model_info(endpoint_config)
-    _maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
+    model_name, model_provider = get_model_info(endpoint_config)
+    maybe_record_budget_cost(store, response, model_name, model_provider, workspace)
     return response
 
 
@@ -1049,8 +877,8 @@ async def gemini_passthrough_generate_content(endpoint_name: str, request: Reque
     response = await traced_passthrough(
         action=PassthroughAction.GEMINI_GENERATE_CONTENT, payload=body, headers=headers
     )
-    model_name, model_provider = _get_model_info(endpoint_config)
-    _maybe_record_budget_cost(
+    model_name, model_provider = get_model_info(endpoint_config)
+    maybe_record_budget_cost(
         store, response, model_name=model_name, model_provider=model_provider, workspace=workspace
     )
     return response
